% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{dc_01_to_noyes}
\alias{dc_01_to_noyes}
\title{Convert 0/1 vector to "No"/"Yes" factor}
\usage{
dc_01_to_noyes(x)
}
\arguments{
\item{x}{Integer or numeric vector. May contain NA. Non-missing values must be 0 or 1.}
}
\value{
A factor with levels c("No", "Yes") corresponding to values 0 and 1,
and the same length as \code{x}.
}
\description{
Convert a vector containing 0 and 1 (and NA) into a factor
with labels "No" and "Yes".
}
\examples{
dc_01_to_noyes(c(0, 1, NA, 1, 0))
}
