\name{focos}
\alias{focos}
\docType{data}
\title{
Focos
}
\description{
Toy data. Numeric vector with record of duration (in hours) of 200 bulbs manufactured with new filament.
}
\usage{data("focos")}

\details{
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(focos)
plot(focos)
}
\keyword{datasets}
