\name{GM}
\alias{GM}
\title{
Geometric Mean
}
\description{
Computes the geometric mean.
}
\usage{
GM(x, na.rm = TRUE)
}
\arguments{
  \item{x}{
a  numeric vector.
}
  \item{na.rm}{
logical value indicating whether \code{NA} values must be removed before calculation. Defaults to \code{FALSE}.
}
}
\details{

}
\value{

}
\references{

}
\author{

}
\note{

}


\seealso{
See also \code{\link{HM}}.
}
\examples{
x = abs(rnorm(100))
GM(x)
}
