\name{HM}
\alias{HM}

\title{
Harmonic mean
}
\description{
Computes the harmonic mean.
}
\usage{
HM(x, na.rm = FALSE)
}

\arguments{
  \item{x}{
a  numeric vector.
}
  \item{na.rm}{
logical value indicating whether \code{NA} values must be removed before calculation. Defaults to \code{FALSE}.
}
}
\details{

}
\value{

}
\references{

}
\author{

}
\note{

}



\seealso{
See also \code{\link{GM}}.
}
\examples{
x = rnorm(100)
HM(x)
}
