\name{particles}
\alias{particles}
\docType{data}
\title{
Particle emissions of Deisel Vehicles
}
\description{
Data consisting of particle emissions measured in 138 diesel vehicles driven at low altitude.
}
\usage{data("particles")}
\format{
Numeric vector of length 138.
}
\details{
All vehicles were manufactured between 1991 and 1996.
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
Navidi, William (2006) Statistics for Engineers and Scientists. McGraw Hill, 1st edition.
}
\examples{
data("particles")
}
\keyword{datasets}
