\name{train}
\alias{train}
\docType{data}
\title{
Train Schedule
}
\description{
Numeric vector containing the arrival time of trian departing from a diptych of the Castelldefels-Barcelona/Sants line collected at the Renfe station.
}
\usage{data("train")}
\format{
Numeric vector of length 91.
}
\details{
Arrival time is in the format hh.mm.
}
\source{
\link{https://www.estadisticaparatodos.es/taller/graficas/tallos_hojas.html}
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(train)
}
\keyword{datasets}
